refactor(tests): point fixtures at oss-litgpt and rework publish script
Use the verified `lightning-ai/oss-litgpt` teamspace for tokenizer/config
fixtures, and rewrite the maintainer-only mirroring script as
tests/publish_fixtures.py, parallelizing repo uploads with litdata.map.
